+++ packages/catalyst/src/cli/commands/build.ts
