+++ packages/catalyst/src/cli/commands/telemetry.ts
