+++ packages/catalyst/src/cli/hooks/telemetry.ts
