+++ packages/catalyst/src/cli/lib/logger.ts
